m3-tests: check properly whether tests failed.
Previously, a successful test reported in the log could mean that it
still failed, because at this point we only checked whether the kernel
shutdown the system.

Thus, I've moved now the actual checking function from the report.py
into a separate python module and use it both from the report.py and
also from the actual test runner.